rename input_builder to key_collector
