+++ components/key_provider/README.md
@@ -1,3 +1,3 @@
 # key_provider component
 
-This is an internal component required by any component that can provide key presses for the `input_builder`.
+This is an internal component required by any component that can provide key presses for the `key_collector`.
