+++ components/matrix_keypad/README.md
@@ -1,7 +1,7 @@
 # keypad component
 
 This component is for matrix keypads.  Define a `keypad` component then add `binary_sensor`s to handle individual keys.  You need to also import the `key_provider` component.
-If you want automatic handling for multiple keys, e.g. PIN entry, use the `input_builder` component.
+If you want automatic handling for multiple keys, e.g. PIN entry, use the `key_collector` component.
 
 The `keys` parameter is optional for the `keypad`, but then you won't be able to check for it in the `binary_sensor`
 and the `input_builder` won't work if you want to use that.
